perf(maintainability) :recycle: Maintainability improvments on multiple files
Modified:
- app/js/appOfficials.js improvments on the code
- index.js removed the updateData function and added the require to the updateData.js file
Added file
- updateData.js Storing the updates functions in a separate file
